Reorganizacion de activity
