+++ app/src/main/res/values/colors.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#FE6D00</color>
-    <color name="colorPrimaryDark">#FE6D00</color>
+    <color name="colorPrimary">#FFFFFF</color>
+    <color name="colorPrimaryDark">#D6D5D3</color>
     <color name="colorAccent">#11B3BA</color>
     <color name="ligh_medim_gray">#F3F3F3</color>
 </resources>
